refactor(core): flatten model/ into single model.rs
The model module (5 files, 573 lines) had no internal coupling between
sub-files. Merge HistoryEntry, MessageBuilder, Model<P>, and context
limits into one file.
